+++ E16/e/src/desktops.h
@@ -1,6 +1,6 @@
 /*
  * Copyright (C) 2000-2007 Carsten Haitzler, Geoff Harrison and various contributors
- * Copyright (C) 2004-2009 Kim Woelders
+ * Copyright (C) 2004-2012 Kim Woelders
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -50,7 +50,7 @@ struct _desk {
 #endif
       Pixmap              pmap;
       Pixmap              pmap_set;
-      unsigned long       pixel;
+      unsigned int        pixel;
       unsigned int        seq_no;
    } bg;
    struct {
