+++ E16/e/src/e16-ecore_hints.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2004-2010 Kim Woelders
+ * Copyright (C) 2004-2012 Kim Woelders
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -37,9 +37,12 @@
 
 int                 ecore_x_client_message32_send(Ecore_X_Window win,
 						  Ecore_X_Atom type,
-						  long mask,
-						  long d0, long d1, long d2,
-						  long d3, long d4);
+						  unsigned int mask,
+						  unsigned int d0,
+						  unsigned int d1,
+						  unsigned int d2,
+						  unsigned int d3,
+						  unsigned int d4);
 
 void                ecore_x_window_prop_card32_set(Ecore_X_Window win,
 						   Ecore_X_Atom atom,
